dtraced: Major refactor into C++.
This commit mostly refactors the job struct into a class, but also
moves around a lot of things relating to jobs into the state class.

This is because the state class is responsible for managing its own
data, including a list of jobs, so it should be there instead of
bundled with job-related things like it was in the C version.

+++ usr.sbin/bhyve/hypertrace.c
@@ -125,7 +125,7 @@ hypertrace_init(struct vmctx *ctx)
 }
 
 static int
-dtraced_sockinit(int *sock, struct vmctx *ctx, uint64_t subs)
+dtraced_sockinit(int *sock, struct vmctx *ctx, unsigned long long subs)
 {
 	size_t l;
 	struct sockaddr_un addr;
@@ -190,14 +190,15 @@ dtraced_sockinit(int *sock, struct vmctx *ctx, uint64_t subs)
 static int
 dtraced_sockinit_rx(struct vmctx *ctx)
 {
-	uint64_t subs = DTD_SUB_ELFWRITE | DTD_SUB_KILL | DTD_SUB_CLEANUP;
+	unsigned long long subs = DTD_SUB_ELFWRITE | DTD_SUB_KILL |
+	    DTD_SUB_CLEANUP;
 	return (dtraced_sockinit(&rx_sockfd, ctx, subs));
 }
 
 static int
 dtraced_sockinit_wx(struct vmctx *ctx)
 {
-	uint64_t subs = DTD_SUB_READDATA;
+	unsigned long long subs = DTD_SUB_READDATA;
 	return (dtraced_sockinit(&wx_sockfd, ctx, subs));
 }
